Refactor code and implement CRUD for clients and users

+++ src/services/clients.service.ts
@@ -1,39 +1,62 @@
-import { Injectable } from '@nestjs/common';
+import { ConflictException, Injectable, InternalServerErrorException, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Clients } from '../entities/clients.entity';
+import { User } from '../entities/user.entity';
+import { CreateClientDto, UpdateClientDto } from '../dtos/clients.dto';
 
 @Injectable()
 export class ClientsService {
   constructor(
     @InjectRepository(Clients)
+    @InjectRepository(User)
     private clientsRepository: Repository<Clients>,
   ) {}
 
-  // Create a new client
-  async create(clientData: Partial<Clients>): Promise<Clients> {
-    const client = this.clientsRepository.create(clientData);
-    return this.clientsRepository.save(client);
+  async create(clientData: CreateClientDto): Promise<Clients> {
+    const existingClient = await this.clientsRepository.findOne({ where: { name: clientData.name } })
+    if (existingClient) {
+      throw new ConflictException('Cliente já está cadastrado.')
+    }
+    try {
+      const client = this.clientsRepository.create(clientData)
+      return await this.clientsRepository.save(client)
+    } catch (error) {
+      throw new InternalServerErrorException('Erro ao criar cliente.')
+    }
   }
 
-  // Get all clients
-  async findAll(): Promise<Clients[]> {
-    return this.clientsRepository.find();
+  async findAll(page: number = 1, limit: number = 10): Promise<{ data: Clients[]; total: number }> {
+    const [data, total] = await this.clientsRepository.findAndCount({
+      skip: (page - 1) * limit,
+      take: limit,
+    })
+    return { data, total };
   }
 
-  // Get one client by ID
   async findOne(id: number): Promise<Clients> {
-    return this.clientsRepository.findOne({ where: { id } });
+    return this.clientsRepository.findOne({ where: { id } })
   }
 
-  // Update a client
-  async update(id: number, clientData: Partial<Clients>): Promise<Clients> {
-    await this.clientsRepository.update(id, clientData);
-    return this.findOne(id);
+  async update(id: number, updateClientDto: UpdateClientDto): Promise<Clients> {
+    const client = await this.findClientByIdOrFail(id)
+    if (!client) throw new NotFoundException('Cliente não encontrado')
+    const updatedClient = this.clientsRepository.merge(client, updateClientDto)
+    return this.clientsRepository.save(updatedClient);
   }
 
-  // Delete a client
   async remove(id: number): Promise<void> {
-    await this.clientsRepository.delete(id);
+    const client = await this.findClientByIdOrFail(id)
+    if (!client) throw new NotFoundException('Cliente não encontrado')
+    await this.clientsRepository.delete(id)
+  } 
+
+  private async findClientByIdOrFail(id: number): Promise<Clients> {
+    const client = await this.clientsRepository.findOne({ where: { id } })
+    if (!client) {
+      throw new NotFoundException('Cliente não encontrado')
+    }
+    return client
   }
 }
+
